Geohash utilities

Various common functions that arise when working often with geohashes

gh_delta(precision)

Arguments

precision

integer precision level desired.

Value

Length-2 numeric vector; the first element is the latitude (y-coordinate) half-width at the input precision, the second element is the longitude (x-coordinate).

Note

Caveat coder: not much is done in the way of consistency checking since this is a convenience function. So e.g. real-valued "precision"s will give results.

References

http://geohash.org/ ( Gustavo Niemeyer's original geohash service )

Author

Michael Chirico

Examples

gh_delta(6)
#> [1] 0.002746582 0.005493164